Renamed generated SOAP client package
Allows for easier integration of other SOAP services provided by AT
