+++ billy-portugal/pom.xml
@@ -289,6 +289,7 @@
 				<version>2.6</version>
 				<executions>
 					<execution>
+						<id>documents</id>
 						<goals>
 							<goal>wsimport</goal>
 						</goals>
@@ -297,7 +298,7 @@
 							<wsdlFiles>
 								<wsdlFile>Fatcorews.wsdl</wsdlFile>
 							</wsdlFiles>
-							<packageName>com.premiumminds.billy.portugal.ws.client</packageName>
+							<packageName>com.premiumminds.billy.portugal.webservices.documents</packageName>
 							<sourceDestDir>src-generated/main/java</sourceDestDir>
 						</configuration>
 					</execution>
