+++ billy-portugal/src/main/java/com/premiumminds/billy/portugal/services/export/webservice/Client.java
@@ -51,21 +51,21 @@ import com.premiumminds.billy.portugal.persistence.entities.PTTaxEntity;
 import com.premiumminds.billy.portugal.services.entities.PTApplication;
 import com.premiumminds.billy.portugal.services.entities.PTGenericInvoice;
 import com.premiumminds.billy.portugal.services.entities.PTRegionContext;
-import com.premiumminds.billy.portugal.ws.client.DateRangeType;
-import com.premiumminds.billy.portugal.ws.client.DebitCreditIndicator;
-import com.premiumminds.billy.portugal.ws.client.DeleteInvoiceRequest;
-import com.premiumminds.billy.portugal.ws.client.DocumentTotals;
-import com.premiumminds.billy.portugal.ws.client.FatcorewsPort;
-import com.premiumminds.billy.portugal.ws.client.FatcorewsPortService;
-import com.premiumminds.billy.portugal.ws.client.InvoiceDataType;
-import com.premiumminds.billy.portugal.ws.client.InvoiceDataType.LineSummary;
-import com.premiumminds.billy.portugal.ws.client.InvoiceStatus;
-import com.premiumminds.billy.portugal.ws.client.InvoiceTypeType;
-import com.premiumminds.billy.portugal.ws.client.ListInvoiceDocumentsType;
-import com.premiumminds.billy.portugal.ws.client.RegisterInvoiceRequest;
-import com.premiumminds.billy.portugal.ws.client.ResponseType;
-import com.premiumminds.billy.portugal.ws.client.Tax;
-import com.premiumminds.billy.portugal.ws.client.TaxType;
+import com.premiumminds.billy.portugal.webservices.documents.DateRangeType;
+import com.premiumminds.billy.portugal.webservices.documents.DebitCreditIndicator;
+import com.premiumminds.billy.portugal.webservices.documents.DeleteInvoiceRequest;
+import com.premiumminds.billy.portugal.webservices.documents.DocumentTotals;
+import com.premiumminds.billy.portugal.webservices.documents.FatcorewsPort;
+import com.premiumminds.billy.portugal.webservices.documents.FatcorewsPortService;
+import com.premiumminds.billy.portugal.webservices.documents.InvoiceDataType;
+import com.premiumminds.billy.portugal.webservices.documents.InvoiceDataType.LineSummary;
+import com.premiumminds.billy.portugal.webservices.documents.InvoiceStatus;
+import com.premiumminds.billy.portugal.webservices.documents.InvoiceTypeType;
+import com.premiumminds.billy.portugal.webservices.documents.ListInvoiceDocumentsType;
+import com.premiumminds.billy.portugal.webservices.documents.RegisterInvoiceRequest;
+import com.premiumminds.billy.portugal.webservices.documents.ResponseType;
+import com.premiumminds.billy.portugal.webservices.documents.Tax;
+import com.premiumminds.billy.portugal.webservices.documents.TaxType;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.commons.lang3.tuple.Pair;
