+++ billy-portugal/src/test/java/com/premiumminds/billy/portugal/services/export/webservice/ClientTest.java
@@ -69,10 +69,10 @@ import com.premiumminds.billy.portugal.services.entities.PTInvoice;
 import com.premiumminds.billy.portugal.services.entities.PTProduct;
 import com.premiumminds.billy.portugal.services.entities.PTTax;
 import com.premiumminds.billy.portugal.util.KeyGenerator;
-import com.premiumminds.billy.portugal.ws.client.DeleteInvoiceResponse;
-import com.premiumminds.billy.portugal.ws.client.InvoiceDataType;
-import com.premiumminds.billy.portugal.ws.client.RegisterInvoiceResponse;
-import com.premiumminds.billy.portugal.ws.client.ResponseType;
+import com.premiumminds.billy.portugal.webservices.documents.DeleteInvoiceResponse;
+import com.premiumminds.billy.portugal.webservices.documents.InvoiceDataType;
+import com.premiumminds.billy.portugal.webservices.documents.RegisterInvoiceResponse;
+import com.premiumminds.billy.portugal.webservices.documents.ResponseType;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
